modify qntz_freq_all to display top1 error as well

+++ matlab/quantize.m
@@ -2,6 +2,6 @@ function xhat = quantize(x,delta)
     if delta == 0
         xhat = x;
     else
-        xhat = delta*(floor(double(x)/delta+0.5));
+        xhat = delta*(floor(x/delta+0.5));
     end
 end
